requirements.txt
altgraph==0.17.4
certifi==2024.7.4
charset-normalizer==3.3.2
huawei-modem-api-client==1.1.5
idna==3.8
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
macholib==1.16.3
modulegraph==0.19.6
more-itertools==10.4.0
py2app==0.28.8
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
python-dotenv==1.0.1
requests==2.32.3
rumps==0.4.0
setuptools==73.0.1
six==1.16.0
typing==3.7.4.3
urllib3==2.2.2
